move datasets to common directory
