+++ src/config/model_config/resnet_test.yaml
@@ -49,7 +49,7 @@ model:
   init_args:
     lr: 0.001
 data:
-  class_path: src.data_modules.cifar10.CIFAR10DataModule
+  class_path: src.datasets.cifar10.CIFAR10DataModule
   init_args:
     batch_size: 64
     num_workers: null
