+++ src/train.py
@@ -10,7 +10,7 @@ from lightning.pytorch.callbacks import (
 )
 from lightning.pytorch.loggers import MLFlowLogger
 
-from src.data_modules.cifar10 import CIFAR10DataModule
+from src.datasets.cifar10 import CIFAR10DataModule
 from src.models.resnet_test import LResNetTest
 from src.utility.style_training_output import model_summary, progress_bar
